+++ packages/ui/components/src/DatePicker/DatePicker.jsx
@@ -2,11 +2,11 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { DatePicker as MuiDatePicker } from '@mui/lab';
 import {
+  DEFAULT_DATE_INPUT_FORMAT,
   DEFAULT_DATE_MASK,
   DEFAULT_INVALID_DATE_MESSAGE,
   DEFAULT_MAX_DATE,
   DEFAULT_MIN_DATE,
-  DEFAULT_VIEW_DATE_FORMAT,
 } from '@astral-frontend/constants';
 import { formatISO, isValid } from 'date-fns';
 
@@ -16,7 +16,7 @@ import DatePickerTextField from './DatePickerTextField';
 const DatePicker = ({
   minDate = DEFAULT_MIN_DATE,
   maxDate = DEFAULT_MAX_DATE,
-  viewFormat = DEFAULT_VIEW_DATE_FORMAT,
+  inputFormat = DEFAULT_DATE_INPUT_FORMAT,
   mask = DEFAULT_DATE_MASK,
   onChange,
   value,
@@ -51,7 +51,7 @@ const DatePicker = ({
         {...props}
         minDate={minDate}
         maxDate={maxDate}
-        inputFormat={viewFormat}
+        inputFormat={inputFormat}
         mask={mask}
         value={value}
         onChange={handleDatePickerChange}
@@ -73,7 +73,7 @@ DatePicker.propTypes = {
   minDate: PropTypes.instanceOf(Date),
   maxDate: PropTypes.instanceOf(Date),
   value: PropTypes.string,
-  viewFormat: PropTypes.string,
+  inputFormat: PropTypes.string,
   mask: PropTypes.string,
   invalidDateMessage: PropTypes.string,
   onChange: PropTypes.func.isRequired,