+++ packages/ui/constants/src/date.js
@@ -1,4 +1,4 @@
-export const DEFAULT_VIEW_DATE_FORMAT = 'dd.MM.yyyy';
+export const DEFAULT_DATE_INPUT_FORMAT = 'dd.MM.yyyy';
 export const DEFAULT_MIN_DATE = new Date('1000-01-01');
 export const DEFAULT_MAX_DATE = new Date('9999-12-31');
 export const DEFAULT_DATE_MASK = '__.__.____';