+++ packages/utils/src/date/date.js
@@ -1,5 +1,5 @@
 import { format } from 'date-fns';
-import { DEFAULT_VIEW_DATE_FORMAT } from '@astral-frontend/constants';
+import { DEFAULT_DATE_INPUT_FORMAT } from '@astral-frontend/constants';
 
 export const formatISOToView = ISODateString =>
-  format(new Date(ISODateString), DEFAULT_VIEW_DATE_FORMAT);
+  format(new Date(ISODateString), DEFAULT_DATE_INPUT_FORMAT);